Fix nightly quiesce logic; minor other updates
Fix the logic when the lights are quiescent at night. The bug was in the
MAXM1 loop where the LED controller was still fired off even though it
was after hours.

There are various other code cleanups, mostly in the debugging code, but
one visible change is that the activity LED only flashes once a minute
during the day rather than flashing the number of minutes every minute.